requirements.txt
streamlit
groq
faiss-cpu
nltk
pypdf
python-docx
python-pptx
pandas
pillow
pytesseract
python-dotenv
sentence-transformers